Add two new custom type hints: StaticAttribute and NonStaticAttribute
These type hints are basically just syntactic sugar for the sets of
types of values that are acceptable as static and non-static attributes
in a PynPoint database.

For now, NonStaticAttribute also encompasses tuples and lists, because
the unit tests sometimes call functions with the "wrong" type. Once the
unit tests have been updated to use only numpy arrays for non-static
attributes, this can be updated accordingly.